# OpenapiSpec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TODO App API",
    "version": "1.0.0",
    "description": "API documentation for a simple TODO app"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },

  "paths": {
    "/todos": {
      "get": {
        "summary": "Get all todos",
        "description": "Returns a list of all todos",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "format": "int64",
                        "description": "Todo ID"
                      },
                      "title": {
                        "type": "string",
                        "description": "Todo title"
                      },
                      "completed": {
                        "type": "boolean",
                        "description": "Todo completion status"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add a new todo",
        "description": "Adds a new todo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Todo title"
                  }
                },
                "required": [
                  "title"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "format": "int64",
                      "description": "Todo ID"
                    },
                    "title": {
                      "type": "string",
                      "description": "Todo title"
                    },
                    "completed": {
                      "type": "boolean",
                      "description": "Todo completion status"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/todos/{id}": {
      "put": {
        "summary": "Update a todo",
        "description": "Updates an existing todo by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Todo ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "Updated todo title"
                  },
                  "completed": {
                    "type": "boolean",
                    "description": "Updated todo completion status"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "format": "int64",
                      "description": "Todo ID"
                    },
                    "title": {
                      "type": "string",
                      "description": "Updated todo title"
                    },
                    "completed": {
                      "type": "boolean",
                      "description": "Updated todo completion status"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a todo",
        "description": "Deletes an existing todo by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Todo ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  }
}